Splitter opp filer og flytter typer der de hører hjemme
